Refactor out GAN functionality more
