+++ satflow/models/gan/__init__.py
@@ -1,1 +1,2 @@
-from .discriminators import GANLoss, PixelDiscriminator, NLayerDiscriminator
+from .discriminators import GANLoss, PixelDiscriminator, NLayerDiscriminator, define_D
+from .generators import define_G
